make history reusable
